Remove legacy initializer action
